Move SymmetricPositiveDefinite to separate module
